clean submit challenge

+++ app/src/main/java/com/example/campusguessr/ChallengeActivity.java
@@ -198,7 +198,7 @@ public class ChallengeActivity extends AppCompatActivity {
                     Map m = (Map<String, Object>) ds.getValue();
                     ObjectMapper mapper = new ObjectMapper();
                     Challenge challenge = mapper.convertValue(m, Challenge.class);
-                    Integer distance = challenge.getLocation().distanceTo(guesses[0]);
+                    Double distance = challenge.getLocation().distanceTo(guesses[0]);
                     long myScore = Math.round((distance * 100.0) / guesses.length);
                     String uId = mAuth.getCurrentUser().getUid();
                     Date currentTime = Calendar.getInstance().getTime();
